Create a GOPS util module and only pass currentTime for gopsSafeToAlignWith
